aded simple country list

+++ lib/api/model/country/country_info.dart
@@ -16,9 +16,9 @@ class CountryInfoBean {
   @JsonKey(name: "country")
   final String country;
   @JsonKey(name: "lat")
-  final int lat;
+  final double lat;
   @JsonKey(name: "long")
-  final int long;
+  final double long;
   @JsonKey(name: "flag")
   final String flag;
